+++ remotefs/posixfs.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/k0sproject/rig/exec"
 	"github.com/k0sproject/rig/log"
-	"github.com/k0sproject/rig/shellescape"
+	"github.com/k0sproject/rig/sh/shellescape"
 )
 
 var (
